allow text alternative to be propagated to the glightbox interfaces
added some code to allow text alternative of img tag to be propagated to the glightbox interfaces by alt attribute or by data-glightbox config.

The glightbox config allow to use a link as a wrapper to an img tag to allow to be focused by the keyboard tabs and propagate the text alternative correctly to the glightbox interface.